=update & delete & filter

+++ app/Http/Controllers/Grades/GradeController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Grades;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreGrades;
+use App\Models\Classroom;
 use App\Models\Grade;
 use CodeZero\UniqueTranslation\UniqueTranslationRule;
 use Illuminate\Http\Request;
@@ -122,9 +123,21 @@ class GradeController extends Controller
      */
     public function destroy(Request $request)
     {
-        $Grades = Grade::findOrFail($request->id)->delete();
-        toastr()->success(trans('messages.Delete'));
-        return redirect()->route('Grades.index');
+        $MyClass_id = Classroom::where('Grade_id', $request->id)->pluck('Grade_id');
+
+        if ($MyClass_id->count() == 0) {
+
+            $Grades = Grade::findOrFail($request->id)->delete();
+            toastr()->success(trans('messages.Delete'));
+            return redirect()->route('Grades.index');
+        } else {
+
+            toastr()->error(trans('Grades_trans.delete_Grade_Error'));
+            return redirect()->route('Grades.index');
+
+        }
+
+
     }
 
 }
